Optimizations
Each expression struct stores the result of the last operation. A new flag for operations determines whether or not the interpreter can recycle the old value if the inputs to the operation are the same. In addition, integer and floats are no longer allocated their own memory and cast to a void pointer in the datavalue struct. Instead, the struct now contains an anonymous union with void *value, int int_value, and double float_value. Also removed the unnecessary set_array command, as a[x]=value has worked for a while now.